# .well-known/openapi.yaml
openapi: 3.0.3
info:
  title: LeeLens API
  description: |
    LeeLens 是一个基于 AI 的代码仓库智能解读平台，能够自动分析任意 GitHub 代码仓库，并生成结构化的项目文档。

    ## AI 友好设计

    本 API 针对以下 AI 工具进行了优化：
    - Claude Code (Anthropic)
    - GitHub Copilot
    - 其他支持 OpenAPI 的 AI 工具

    ## 核心功能

    - **仓库管理**：添加、删除、查看代码仓库
    - **任务管理**：执行、监控、重试分析任务
    - **文档管理**：查看、编辑、导出生成的文档
    - **API Key 管理**：配置多个 LLM API Key，支持自动切换
    - **数据同步**：与其他 LeeLens 实例同步数据
    - **用户需求管理**：提交和跟踪用户需求

    ## 响应格式

    所有 API 响应遵循统一格式：

    ### 成功响应
    ```json
    {
      "code": 0,
      "message": "success",
      "data": { ... }
    }
    ```

    ### 错误响应
    ```json
    {
      "error": "error message"
    }
    ```
  version: 1.0.0
  contact:
    name: LeeLens
    url: https://gitee.com/li-yuyanglee/eino-lee-lens
servers:
  - url: http://localhost:8080
    description: 开发环境
  - url: https://leelens.fly.dev
    description: 生产环境
...
paths:
...
  /api/sync:
    post:
      tags:
        - sync
      summary: 开始推送同步
      description: 将本地数据同步到目标服务器
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSyncRequest'
      responses:
        '200':
          description: 同步已启动
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartSyncResponse'
  /api/sync/pull:
    post:
      tags:
        - sync
      summary: 开始拉取同步
      description: 从目标服务器拉取数据
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PullSyncRequest'
      responses:
        '200':
          description: 拉取已启动
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartSyncResponse'
  /api/sync/status/{sync_id}:
    get:
      tags:
        - sync
      summary: 获取同步状态
      description: 查询同步任务的状态和进度
      parameters:
        - name: sync_id
          in: path
          required: true
          description: 同步任务 ID
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncStatus'
  /api/sync/event-list:
    get:
      tags:
        - sync
      summary: 获取同步事件列表
      description: 获取同步历史记录
      parameters:
        - name: repository_id
          in: query
          description: 仓库 ID
          schema:
            type: integer
        - name: mode
          in: query
          description: 同步模式
          schema:
            type: string
            enum: [push, pull, all]
        - name: limit
          in: query
          description: 返回记录数量
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: OK
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SyncEvent'
...
  /api/sync/document-list:
    get:
      tags:
        - sync
      summary: 获取文档列表
      description: 获取仓库的文档列表
      parameters:
        - name: repository_id
          in: query
          required: true
          description: 仓库 ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: OK
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SyncDocument'
...
  /api/sync/target-save:
    post:
      tags:
        - sync
      summary: 保存同步目标
      description: 添加或更新同步目标服务器
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: 目标服务器 URL
              required:
                - url
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncTargetResponse'
  /api/sync/target-delete:
    post:
      tags:
        - sync
      summary: 删除同步目标
      description: 删除同步目标服务器
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  description: 同步目标 ID
              required:
                - id
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncTargetDeleteResponse'
  /api/sync/pull-export:
    post:
      tags:
        - sync
      summary: 生成拉取导出数据
      description: 生成用于拉取同步的数据包
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repository_id:
                  type: integer
                  description: 仓库 ID
                document_ids:
                  type: array
                  items:
                    type: integer
                  description: 文档 ID 列表
              required:
                - repository_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullExportResponse'
  /api/sync/repository-upsert:
    post:
      tags:
        - sync
      summary: 创建或更新仓库
      description: 在目标服务器上创建或更新仓库信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepositoryUpsertRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryUpsertResponse'
  /api/sync/repository-clear:
    post:
      tags:
        - sync
      summary: 清空仓库数据
      description: 清空仓库的所有文档和任务数据
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repository_id:
                  type: integer
                  description: 仓库 ID
              required:
                - repository_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryClearResponse'
  /api/sync/task-create:
    post:
      tags:
        - sync
      summary: 创建任务
      description: 在目标服务器上创建任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreateResponse'
  /api/sync/document-create:
    post:
      tags:
        - sync
      summary: 创建文档
      description: 在目标服务器上创建文档
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentCreateResponse'
  /api/sync/task-update-docid:
    post:
      tags:
        - sync
      summary: 更新任务的文档 ID
      description: 将任务与文档关联
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                task_id:
                  type: integer
                  description: 任务 ID
                document_id:
                  type: integer
                  description: 文档 ID
              required:
                - task_id
                - document_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskUpdateDocIDResponse'
  /api/sync/task-usage-create:
    post:
      tags:
        - sync
      summary: 创建任务用量记录
      description: 在目标服务器上创建或覆盖任务用量记录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUsageCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskUsageCreateResponse'
...
components:
...
  schemas:
...
    StartSyncRequest:
      type: object
      properties:
        target_server:
          type: string
          description: 目标服务器地址
        repository_id:
          type: integer
          description: 仓库 ID
        document_ids:
          type: array
          items:
            type: integer
          description: 文档 ID 列表
        clear_target:
          type: boolean
          description: 是否清空目标服务器数据
      required:
        - target_server
        - repository_id
    PullSyncRequest:
      type: object
      properties:
        target_server:
          type: string
          description: 目标服务器地址
        repository_id:
          type: integer
          description: 仓库 ID
        document_ids:
          type: array
          items:
            type: integer
          description: 文档 ID 列表
        clear_local:
          type: boolean
          description: 是否清空本地数据
      required:
        - target_server
        - repository_id
    StartSyncResponse:
      type: object
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          properties:
            sync_id:
              type: string
              description: 同步任务 ID
            repository_id:
              type: integer
              description: 仓库 ID
            total_tasks:
              type: integer
              description: 总任务数
            status:
              type: string
              description: 状态
    SyncStatus:
      type: object
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          properties:
            sync_id:
              type: string
            repository_id:
              type: integer
            total_tasks:
              type: integer
            completed_tasks:
              type: integer
            failed_tasks:
              type: integer
            status:
              type: string
            current_task:
              type: string
            started_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    SyncEvent:
      type: object
      properties:
        id:
          type: integer
        repository_id:
          type: integer
        mode:
          type: string
          enum: [push, pull]
        status:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...
    SyncDocument:
      type: object
      properties:
        id:
          type: integer
        repository_id:
          type: integer
        task_id:
          type: integer
        title:
          type: string
        updated_at:
          type: string
          format: date-time
...
    SyncTargetResponse:
      type: object
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          properties:
            id:
              type: integer
            url:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    SyncTargetDeleteResponse:
      type: object
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          properties:
            id:
              type: integer
    PullExportResponse:
      type: object
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          description: 导出数据
    RepositoryUpsertRequest:
      type: object
      properties:
        id:
          type: integer
          description: 仓库 ID（更新时使用）
        name:
          type: string
          description: 仓库名称
        url:
          type: string
          description: 仓库 URL
        status:
          type: string
          description: 状态
        branch:
          type: string
          description: 分支
    RepositoryUpsertResponse:
      type: object
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          properties:
            repository_id:
              type: integer
            name:
              type: string
    RepositoryClearResponse:
      type: object
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          properties:
            repository_id:
              type: integer
    TaskCreateRequest:
      type: object
      properties:
        task_id:
          type: integer
          description: 任务 ID
        repository_id:
          type: integer
          description: 仓库 ID
        title:
          type: string
          description: 任务标题
        status:
          type: string
          description: 状态
        writer_name:
          type: string
          description: 写入器名称
        sort_order:
          type: integer
          description: 排序顺序
    TaskCreateResponse:
      type: object
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          properties:
            task_id:
              type: integer
            repository_id:
              type: integer
            title:
              type: string
    DocumentCreateRequest:
      type: object
      properties:
        id:
          type: integer
          description: 文档 ID
        repository_id:
          type: integer
          description: 仓库 ID
        task_id:
          type: integer
          description: 任务 ID
        title:
          type: string
          description: 标题
        content:
          type: string
          description: 内容
        version:
          type: integer
          description: 版本号
    DocumentCreateResponse:
      type: object
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          properties:
            document_id:
              type: integer
            repository_id:
              type: integer
            task_id:
              type: integer
    TaskUpdateDocIDResponse:
      type: object
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          properties:
            task_id:
              type: integer
            document_id:
              type: integer
    TaskUsageCreateRequest:
      type: object
      properties:
        task_id:
          type: integer
          description: 任务 ID
        repository_id:
          type: integer
          description: 仓库 ID
        input_tokens:
          type: integer
          description: 输入 Token 数
        output_tokens:
          type: integer
          description: 输出 Token 数
        total_tokens:
          type: integer
          description: 总 Token 数
        duration_ms:
          type: integer
          description: 执行时长（毫秒）
    TaskUsageCreateResponse:
      type: object
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          properties:
            task_id:
              type: integer
...